+++ app-wearos/src/main/java/de/danoeh/antennapod/wearos/sync/WearDataRepository.kt
@@ -1,10 +1,18 @@
 package de.danoeh.antennapod.wearos.sync
 
+import android.content.Context
+import android.util.Log
+import com.google.android.gms.wearable.Wearable
 import de.danoeh.antennapod.model.feed.Feed
 import de.danoeh.antennapod.model.feed.FeedItem
+import de.danoeh.antennapod.net.sync.wearinterface.WearConnectionUtils
 import de.danoeh.antennapod.net.sync.wearinterface.WearNowPlaying
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.tasks.await
 
 /**
  * Singleton holding the last data received from the phone for each data path.
@@ -31,5 +39,24 @@ object WearDataRepository {
 
     fun updateFeeds(path: String, feeds: List<Feed>) {
         _feedsByPath.value += (path to feeds)
+    }
+
+    fun requestDataFromPhone(context: Context, scope: CoroutineScope, path: String, tag: String) {
+        scope.launch(Dispatchers.IO) {
+            val supported = WearConnectionUtils.isPhoneSupported(context)
+            if (!supported) {
+                return@launch
+            }
+            try {
+                val nodes = Wearable.getNodeClient(context).connectedNodes.await()
+                val node = nodes.firstOrNull() ?: run {
+                    Log.w(tag, "No connected nodes")
+                    return@launch
+                }
+                Wearable.getMessageClient(context).sendMessage(node.id, path, null).await()
+            } catch (e: Exception) {
+                Log.e(tag, "Failed to send message to phone", e)
+            }
+        }
     }
 }
